Overview and collection collapse belong to front content
